使用UniqueSpiderScheduler避免重复抓取

+++ src/main/java/com/geccocrawler/gecco/request/HttpRequest.java
@@ -36,7 +36,7 @@ public interface HttpRequest extends Cloneable {
 	
 	public void clearCookie();
 
-	public int getPriority();
+	public long getPriority();
 	
-	public void setPriority(int prio);
+	public void setPriority(long prio);
 }
